Added skeleton for towers.
Trying to make use of the data structures we talked about. Matrix for
grid. Tower array for all towers. Enemy array for all enemies. When the
game updates, it should update each of these arrays. When the game
draws, it should draw everything in these arrays. I will try to add a
simple attack animation for the towers next. For this I will need to add
the ogre to the list and update the draw method for the towers.